Implementing Example project, but... needs rename
I found that the name of the built framework is changed to `youtube_ios_player_helper` when framework (clang module) is enabled because of the clang module specification:
https://github.com/CocoaPods/Core/issues/205

To make things worse this only happens when clang module is enabled, so make thing so much completecated. I can't keep using the name `youtube-ios-player-helper` anymore because of this, so I'm going to rename the entire module name to simple `YTPlayerView` in next commit.