Added a debug interface to WK.
This might have to be removed for production build settings, but I believe this won't cause any problems as long as we don't call this from HTML. Even so it just emits a line of log and that's it.

Also I learned how to use JS user script interface so I think we can use this as our advantege to callback events from JS to Obj-C wrapper so that we no longer have to use ytplayer:// hacky thing.